fix: 프래그먼트 꺼진 후에도 남아있는 데이터 처리

+++ app/src/main/java/com/project/meongcare/symptom/model/entities/SymptomData.kt
@@ -8,7 +8,7 @@ data class ResultSymptom(
 )
 
 data class ToAddSymptom(
-    val dogId: Long,
+    val dogId: Int,
     val symptomString: String,
     val note: String,
     val dateTime: String,
